+++ TOMB4/game/trng/trng_test_position.h
@@ -9,7 +9,7 @@ enum TestPositionFlags {
 	TPOS_OPPOSITE_FACING = 0x0010,
 };
 
-long NGTestLaraDistance(PHD_VECTOR* target, ITEM_INFO* item, ITEM_INFO* l);
-PHD_VECTOR NGCalculatePositionForTestPosition(NG_TEST_POSITION* test_position);
-bool NGTestLaraPosition(NG_TEST_POSITION* test_position, ITEM_INFO* item, ITEM_INFO* l);
+int32_t NGTestLaraDistance(PHD_VECTOR *target, ITEM_INFO *item, ITEM_INFO *l);
+PHD_VECTOR NGCalculatePositionForTestPosition(NG_TEST_POSITION *test_position);
+bool NGTestLaraPosition(NG_TEST_POSITION *test_position, ITEM_INFO *item, ITEM_INFO *l);
 bool NGMoveLara(bool set_stop_animation_when_goal_reached);
